✨ feat(brands): refactor brand list components and add loading skeleton
