+++ src/app/brands/_components/brand-list.tsx
